pyproject.toml
[project]
name = "nexusgate"
version = "1.0.0"
description = "Industrial-grade unified API gateway for databases and file storage."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.1",
    "watchfiles>=0.21.0",
    "cachetools>=5.3.3",
    "structlog>=24.1.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    "zipstream-ng>=1.7.1",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "sqlglot>=23.0.0",
    "SQLAlchemy[asyncio]>=2.0.28",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0"
]

[project.optional-dependencies]
redis = ["redis[hiredis]>=5.0.0"]
image = ["Pillow>=10.2.0"]
mssql = ["aioodbc>=0.5.0", "pyodbc>=5.1.0"]
oracle = ["cx_Oracle>=8.3.0"]

[project.scripts]
nexusgate = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
